Introducing new KeyProviderRequirementsBase class
This commit also fixes an issue in the
PrivateFileKeyProvider which should not have tried
to read the file when the key is new and unsaved.
